+++ public/ejs/namegame/group/wait.ejs
@@ -3,5 +3,5 @@
     <p class="text-center mt-3">みなさんの準備が<br />できるまでお待ち下さい</p>
     <ul class="list-group"></ul>
 </div>
-<script src="../../../js/DiceGame/waitClient.js" type="module"></script>
+<script src="../../../js/nameGame/waitClient.js" type="module"></script>
 <%- include( '../_footer.ejs' ); %>
