+++ public/index.html
@@ -133,7 +133,7 @@
 
                 <div class="col-6 col-md-3 col-lg-2 col-xl-2 mt-2">
                     <div class="card home-card">
-                        <a href="./html/namegame/game/dev.html">
+                        <a href="./html/namegame/group/group.html">
                             <div class="m-3">
                                 <img
                                     src="./images/NameGame.svg"
